# app-mod/money-backend-app/openapi.yaml
openapi: 3.1.0
...
info:
  title: money-log 백엔드 API
  version: v1
  description: |
    가계부 백엔드 REST API. 50개 오퍼레이션 / 34개 경로.

    ## 응답 규격 — 실패도 HTTP 200 이다

    성공·실패 모두 `{ "resCode": <숫자>, "data": { ... } }` 한 형태다(헌장 원칙 III).
    성공은 `resCode: 200`, 비즈니스·검증 실패는 **정수 4자리** 에러코드다.

    **HTTP 상태만 보고 성공으로 읽으면 안 된다.** 아래 둘을 제외한 모든 API 가
    실패에도 HTTP 200 을 돌려준다.

    | 예외 | 이유 |
    |---|---|
    | `GET /api/v1/expend-groups/icons/{filename}` | `<img>`/fetch 로 받는 이미지라 성공·실패 모두 래퍼가 아니다. 실패는 본문 없는 HTTP 상태코드다 |
    | `GET /api/v1/expense-incomes/excel/template` | **성공만** 예외다(.xlsx 바이너리). 실패는 래퍼 + HTTP 200 이다 — 사용자가 다운로드 버튼을 누르는 흐름이라 사유를 화면에 띄워야 한다 |

    실패 응답의 `data` 는 `{ "message": "..." }` 한 칸이며, 엑셀 행 검증 실패(`3502`)만
    `errors[]` 를 더 싣는다.

    ## HTTP 메서드

    GET 조회 · POST 생성·로그인 · PATCH 수정 · DELETE 삭제. **PUT 은 쓰지 않는다.**

    PATCH 는 **omit = 유지**다. 보내지 않은 필드는 그대로 두고 보낸 필드만 바꾼다.
    그래서 요청 스키마의 모든 필드가 선택이며, 명시적 `null` 과 omit 을 구분한다.

    ## 인증

    로그인이 주는 `accessToken` 을 `Authorization: Bearer <token>` 으로 보낸다.
    Access 1일 · Refresh 7일이고 **회원당 활성 세션은 1건**이라 새 로그인이 기존 세션을
    폐기한다. 검증은 서명·만료뿐 아니라 **DB 세션 대조까지** 한다.

    권한은 관리자 `1`, 일반 `3`(가입 기본값). 사용자는 **본인 데이터만** 접근한다.

    ## 소유자 판정과 에러코드

    "없음"과 "타인 소유"를 **같은 코드**로 낸다. 나누면 ID 를 훑는 것만으로 남의 데이터가
    존재한다는 사실이 새어 나간다.
servers:
  - url: http://localhost:8081
    description: 로컬 개발 서버
...
paths:
...
  /api/v1/expend-targets:
    get:
      tags: [목표금액]
      summary: "5.1 목표금액 목록"
      description: |
        **모집단이 목표 행이 아니라 사용 중 지출유형이다**(FR-509). 목표를 한 번도 정하지
        않은 유형도 `defaultTargetAmount=0`·`monthlyTargetAmount=null` 로 한 줄 나온다 —
        목표금액 화면이 "아직 안 정한 유형"을 보여주고 거기에 값을 넣게 해야 한다.

        **006 의 여섯 API 중 페이징이 있는 것은 여기뿐이다.** 통계 응답의 배열 3종은 한
        통계 객체의 구성 요소라 `data.list` 규칙의 대상이 아니다.

        ### 두 실패 코드가 다르다

        ```text
        year·month 누락·범위 오류    → 3603   (자원의 좌표)
        offset·limit 누락·범위·배수  → 9001   (요청 형식)
        ```

        검증 순서도 **연·월이 먼저**다 — 둘 다 어긋난 요청은 `3603` 이다.

        **`totalCount` 도 사용 중 유형만 센다**(FR-526). 목록만 필터하고 건수를 필터하지
        않으면 페이지 수가 맞지 않아 화면이 빈 마지막 페이지를 그린다.

        **삭제 표시(`deleted=true`)는 조건이 아니다.** 보는 것은 `in_use` 뿐이라, 사용
        중인 채로 삭제 표시만 된 유형은 목록에 **남는다** — 그래야 FR-511 이 지키려는
        목표 행을 화면에서 읽을 수 있다.
      parameters:
        - $ref: "#/components/parameters/TargetYear"
        - $ref: "#/components/parameters/TargetMonth"
        - { $ref: "#/components/parameters/Offset" }
        - { $ref: "#/components/parameters/Limit" }
        - name: expendGroupName
          in: query
          required: false
          schema: { type: string }
          description: "지출유형 이름 **부분 일치** 검색. 사용 중 유형 범위 안에서 동작한다"
      responses:
        "200":
          description: "실패: `3603` 연·월 · `9001` 페이징"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - properties:
                      data: { $ref: "#/components/schemas/ExpendTargetListResponse" }
  /api/v1/expend-targets/{year}/{month}/{expendGroupId}:
    get:
      tags: [목표금액]
      summary: "5.2 목표금액 상세 조회"
      description: |
        ### 판정 순서가 결과 코드를 바꾼다

        ```text
        1. year·month 범위                        오류 → 3603
        2. expendGroupId 조회 (본인 소유?)   없음·타인 → 3103
        3. in_use = false 인가                        → 3601
        4. 기본·월별을 각각 읽어 응답
        ```

        **2번이 3번보다 먼저다.** 남의 유형 ID 로 접근했는데 그게 마침 `in_use=false` 라면
        `3601` 이 아니라 `3103` 이어야 한다 — `3601` 을 내는 순간 그 ID 가 실재한다는
        사실이 코드 차이로 새어 나간다.

        **목록은 조용히 빼는데 단건은 `3601` 로 거절한다.** 단건은 사용자가 그 유형을
        지목한 것이라 왜 볼 수 없는지 알려줘야 하고, 목록은 지목한 것이 아니라 알릴
        대상이 없다.

        **삭제 표시된 유형도 읽힌다.** 조건은 `in_use` 뿐이다(FR-511).
      parameters:
        - $ref: "#/components/parameters/TargetYearPath"
        - $ref: "#/components/parameters/TargetMonthPath"
        - $ref: "#/components/parameters/ExpendGroupIdPath"
      responses:
        "200":
          description: "실패: `3603` 연·월 · `3103` 없음·타인 · `3601` 미사용 유형"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - properties:
                      data: { $ref: "#/components/schemas/ExpendTargetDetailResponse" }
  /api/v1/expend-targets/default/{expendGroupId}:
    patch:
      tags: [목표금액]
      summary: "5.3 기본 목표금액 등록·수정"
      description: |
        **`PATCH` 이고 upsert 다**(FR-512). 행이 없으면 만들고 있으면 갱신한다 —
        사용자에게 "최초 설정"과 "변경"은 같은 행위(목표금액 칸에 숫자를 넣는 것)라
        API 를 나누면 프론트가 "이미 있는가"를 먼저 조회해 분기해야 한다.
        `PUT` 은 이 저장소에서 쓰지 않는다.

        ### 판정 순서

        ```text
        1. expendGroupId 조회 (본인 소유?)   없음·타인 → 3103
        2. in_use = false 인가                        → 3601
        3. targetAmount 가 0 ~ 100,000,000 밖         → 3602
        4. upsert (INSERT ... ON CONFLICT DO UPDATE)
        ```

        **금액이 가장 나중이다.** 순서가 뒤집히면 남의 유형에 범위 밖 금액을 보냈을 때
        `3103` 대신 `3602` 가 나가 그 ID 가 실재함이 드러난다.

        **동시 요청에서도 둘 다 200 이다.** "조회해서 없으면 INSERT" 로는 두 요청이 같은
        순간 "없음"을 보고 하나가 `ux_target_default` 위반으로 `9000` 을 낸다.

        **월별 목표를 건드리지 않는다**(FR-505). 두 층은 독립이며 기본을 월별로 복사하지도
        않는다 — 복사하면 "이 달은 기본을 따른다"와 "마침 기본과 같은 값으로 정했다"가
        구분되지 않는다.

        **몸통 필드 이름이 5.4 와 다르다** — 여기는 `defaultTargetAmount` 다.
      parameters:
        - $ref: "#/components/parameters/ExpendGroupIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ExpendTargetDefaultUpsertRequest" }
      responses:
        "200":
          description: |
            성공 `data` 는 **담당한 층만** 싣는다 — 월별 목표는 이 호출로 변하지 않는다.

            실패: `3103` 없음·타인 · `3601` 미사용 유형 · `3602` 금액 누락·범위 밖
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - properties:
                      data: { $ref: "#/components/schemas/ExpendTargetDefaultResponse" }
  /api/v1/expend-targets/monthly/{year}/{month}/{expendGroupId}:
    patch:
      tags: [목표금액]
      summary: "5.4 월별 목표금액 등록·수정"
      description: |
        규칙은 5.3 과 같고 좌표에 연·월이 붙어 판정 맨 앞에 `3603` 이 하나 더 붙는다.
        **기본 목표를 건드리지 않는다**(FR-505).

        ```text
        1. year·month 범위                        오류 → 3603
        2. expendGroupId 조회 (본인 소유?)   없음·타인 → 3103
        3. in_use = false 인가                        → 3601
        4. targetAmount 가 0 ~ 100,000,000 밖         → 3602
        5. upsert (INSERT ... ON CONFLICT DO UPDATE)
        ```

        **`0` 을 그대로 저장한다.** "그 달엔 쓰지 않겠다"라 행이 없는 상태(`null`)와 다르며
        (FR-506), 0 을 "지우기"로 해석해 DELETE 로 바꾸면 그 구분이 사라진다.

        **몸통 필드 이름이 5.3 과 다르다** — 여기는 `monthlyTargetAmount` 다.

        5.2 와 경로의 연·월 자리가 겹쳐 보이지만 이쪽은 `/monthly` 하위라 갈리고 메서드도
        다르다.
      parameters:
        - $ref: "#/components/parameters/TargetYearPath"
        - $ref: "#/components/parameters/TargetMonthPath"
        - $ref: "#/components/parameters/ExpendGroupIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ExpendTargetMonthlyUpsertRequest" }
      responses:
        "200":
          description: |
            성공 `data` 는 **담당한 층만** 싣는다 — 기본 목표는 이 호출로 변하지 않는다.

            실패: `3603` 연·월 · `3103` 없음·타인 · `3601` 미사용 유형 · `3602` 금액
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - properties:
                      data: { $ref: "#/components/schemas/ExpendTargetMonthlyResponse" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 로그인 응답의 `accessToken`
  parameters:
    Offset:
      name: offset
      in: query
      required: true
      schema: { type: integer, minimum: 0 }
      description: "건너뛸 건수. 0 이상이고 `limit` 의 **배수**. 빠지면 `9001`"
    Limit:
      name: limit
      in: query
      required: true
      schema: { type: integer, minimum: 1 }
      description: "가져올 건수. 1 이상. 빠지면 `9001`"
    TargetYear:
      name: year
      in: query
      required: true
      schema: { type: integer, minimum: 2000, maximum: 2100 }
      description: |
        조회 연도. **2000~2100** 이며 벗어나거나 빠지면 `3603`(페이징의 `9001` 이 아니다).

        `ck_statistics_month` 는 월만 검사하므로 DB 가 대신 막아주지 않는다 — 연도 판정은
        애플리케이션 몫이다. **현재 연도 기준 상대 범위를 쓰지 않는 것**은 경계가 해마다
        움직이면 경계 시험이 시간에 의존하기 때문이다
    TargetMonth:
      name: month
      in: query
      required: true
      schema: { type: integer, minimum: 1, maximum: 12 }
      description: "조회 월. 벗어나거나 빠지면 `3603`"
    TargetYearPath:
      name: year
      in: path
      required: true
      schema: { type: integer, minimum: 2000, maximum: 2100 }
      description: "연도. **2000~2100**, 벗어나면 `3603`"
    TargetMonthPath:
      name: month
      in: path
      required: true
      schema: { type: integer, minimum: 1, maximum: 12 }
      description: "월. 벗어나면 `3603`"
    ExpendGroupIdPath:
      name: expendGroupId
      in: path
      required: true
      schema: { type: integer, format: int64 }
      description: "지출유형 PK. 없거나 타인 소유면 `3103`(미사용의 `3601` 보다 **먼저** 판정)"
  schemas:
...
    ExpendTargetDefaultUpsertRequest:
      type: object
      description: |
        5.3 의 몸통. **필드 이름이 5.4 와 다르다** — 두 층이 독립이라 요청에서도 어느 층을
        건드리는지가 이름으로 드러난다.
      required: [defaultTargetAmount]
      properties:
        defaultTargetAmount:
          type: integer
          format: int64
          minimum: 0
          maximum: 100000000
          description: |
            월 단위 기본 목표금액. **빠뜨려도 `3602`** 다 — 필드가 하나뿐이라 "빠뜨렸다"와
            "값이 틀렸다"를 가르는 것이 사용자에게 주는 정보가 없다
    ExpendTargetMonthlyUpsertRequest:
      type: object
      description: 5.4 의 몸통. 이름 규칙은 `ExpendTargetDefaultUpsertRequest` 와 같다
      required: [monthlyTargetAmount]
      properties:
        monthlyTargetAmount:
          type: integer
          format: int64
          minimum: 0
          maximum: 100000000
          description: "그 연·월의 목표금액. **`0` 은 \"그 달엔 쓰지 않겠다\"** 라 유효한 값이다"
    ExpendTargetResponse:
      type: object
      description: |
        5.1 목록의 한 줄. **적용 금액 필드는 없다** — 화면이
        `monthlyTargetAmount ?? defaultTargetAmount` 로 낸다. 서버가 같은 정보를 두 형태로
        내리면 둘이 어긋날 여지만 생긴다.
      properties:
        expendGroupId: { type: integer, format: int64 }
        expendGroupName:
          type: string
          description: "지출유형의 **현재** 이름. 목표금액은 지금 유효한 설정이라 원본에서 읽는다"
        defaultTargetAmount:
          type: integer
          format: int64
          description: "기본 목표. **행이 없으면 `0`** 이다 — `null` 이 아니다(FR-507)"
        monthlyTargetAmount:
          type: integer
          format: int64
          nullable: true
          description: |
            그 연·월에 **별도 저장된** 목표. **행이 없으면 `null`**(기본값을 쓴다),
            0 으로 저장했으면 `0`(그 달엔 쓰지 않는다). 둘은 다른 상태다(FR-506).

            **`null` 일 때 필드를 생략하지 않는다** — 프론트가
            `'monthlyTargetAmount' in obj` 로 분기하면 생략과 `null` 이 다른 결과를 낸다
    ExpendTargetListResponse:
      type: object
      description: "부가 필드가 `list` 와 **같은 레벨**이다(FR-526)"
      properties:
        year:
          type: integer
          description: "요청 연도. 되돌려주므로 화면이 응답 하나로 표를 그릴 수 있다"
        month: { type: integer }
        list:
          type: array
          items: { $ref: "#/components/schemas/ExpendTargetResponse" }
          description: "**사용 중** 유형만"
        offset: { type: integer }
        limit: { type: integer }
        totalCount:
          type: integer
          format: int64
          description: "**사용 중 유형만** 센다(FR-526). 목록과 모집단이 같아야 페이지 수가 맞는다"
    ExpendTargetDetailResponse:
      type: object
      description: "5.2. 금액 두 필드는 `ExpendTargetResponse` 와 같고 연·월이 앞에 붙는다"
      properties:
        year: { type: integer }
        month: { type: integer }
        expendGroupId: { type: integer, format: int64 }
        expendGroupName: { type: string, description: "**현재** 이름" }
        defaultTargetAmount:
          type: integer
          format: int64
          description: "행이 없으면 `0`"
        monthlyTargetAmount:
          type: integer
          format: int64
          nullable: true
          description: "행이 없으면 `null`. 생략하지 않는다"
    ExpendTargetDefaultResponse:
      type: object
      description: |
        5.3 의 결과. **담당한 층만** 싣는다 — 월별 목표는 이 호출로 변하지 않으므로
        여기 실을 이유가 없고, 함께 실으면 "이 호출이 월별도 건드렸나"라는 오해를 만든다.
      properties:
        expendGroupId: { type: integer, format: int64 }
        expendGroupName: { type: string, description: "**현재** 이름" }
        defaultTargetAmount:
          type: integer
          format: int64
          description: 저장된 월 단위 기본 목표금액
    ExpendTargetMonthlyResponse:
      type: object
      description: "5.4 의 결과. 연·월이 붙는 것 말고는 `ExpendTargetDefaultResponse` 와 같다"
      properties:
        year: { type: integer }
        month: { type: integer }
        expendGroupId: { type: integer, format: int64 }
        expendGroupName: { type: string, description: "**현재** 이름" }
        monthlyTargetAmount:
          type: integer
          format: int64
          description: |
            저장된 월별 목표금액. **여기서는 `null` 이 될 수 없다** — 방금 저장한 값이라
            "행이 없음"이 나올 수 없고, `0` 도 정상 결과다


    # ── 응답 래퍼 ──────────────────────────────────────────────────────────
    Envelope:
      type: object
      description: |
        모든 응답의 겉모양. **최상위 필드는 `resCode`·`data` 둘뿐이다.**
      required: [resCode, data]
      properties:
        resCode:
          type: integer
          description: "성공 `200`. 실패는 **정수 4자리** 에러코드"
          example: 200
        data:
          type: object
...
security:
  - bearerAuth: []